\name{NLSdata}
\alias{NLSdata}
\alias{NLSData}
\alias{nlsdata}
\title{NLSdata: A Package for National Longitudinal Survey Analysis}

\description{
Welcome to NLSdata, a package facilitating analysis of the National Longitudinal Surveys
via the NLS Investigator. Instructions for getting started can be found at
\url{http://blog.analyzestuff.com/2014/02/nlsdata-r-package-for-national.html}.

\preformatted{Furthermore, help pages are available for the following functions:
CreateNLSdata
CreateTimeSeriesDf
ThrowAwayDataForBalance
RosterToLongDf
summary.NLSdata
}
}
